# ws/docs/openapi/provisioning.openapi.yaml
openapi: 3.1.0
info:
  title: Sukko Provisioning API
  version: 1.0.0
  description: |
    REST API for managing tenants, JWT signing keys, API keys, routing rules,
    channel rules, and quotas in the Sukko multi-tenant WebSocket platform.

    ## Authentication

    All `/api/v1/` endpoints require authentication via one of:
    - **Admin JWT**: `Authorization: Bearer {admin_jwt}` — Ed25519 keypair auth with `iss: "sukko-admin"`. Checked first; if present and valid, bypasses tenant JWT.
    - **Tenant JWT**: `Authorization: Bearer {jwt}` — validated against per-tenant asymmetric keys (ES256, RS256, EdDSA).

    Health, readiness, version, edition, config, and metrics endpoints (at root level, not under `/api/v1/`) require no authentication.

    ## Authorization

    Role-based access control with tenant isolation:
    - **admin/system**: Full access to all tenants and operations
    - **tenant owner**: Access to own tenant's resources only (JWT `tenant_id` must match)

    Operations like creating tenants, suspending, and modifying routing rules require `admin` or `system` role.

    ## Pagination

    All list endpoints support `?limit=N&offset=M` with defaults `limit=50, offset=0` (max limit=100).
    Responses include `{ items, total, limit, offset }`.

    ## Error Format

    All errors return: `{"code": "UPPER_SNAKE_CASE", "message": "human-readable"}`.
  contact:
    name: Sukko Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/tenants/{tenantSlug}/channel-rules:
    parameters:
      - $ref: "#/components/parameters/tenantSlug"

    get:
      operationId: getChannelRules
      summary: Get channel rules
      description: Returns channel access control rules for the tenant.
      tags: [Channel Rules]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Channel rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantChannelRules"
        "404":
          description: No channel rules configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: setChannelRules
      summary: Set channel rules
      description: |
        Replaces channel access control rules for the tenant. Requires admin or system role.
        Patterns support wildcards (`*`) and placeholders (`{subject}`, `{group_id}`).
      tags: [Channel Rules]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetChannelRulesRequest"
      responses:
        "200":
          description: Rules set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantChannelRules"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteChannelRules
      summary: Delete channel rules
      description: Removes channel rules for the tenant. Falls back to global patterns. Requires admin or system role.
      tags: [Channel Rules]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rules deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: deleted
        "404":
          description: No rules configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/tenants/{tenantSlug}/test-access:
    parameters:
      - $ref: "#/components/parameters/tenantSlug"

    post:
      operationId: testAccess
      summary: Test channel access
      description: Computes which channel patterns would be allowed for given groups. Useful for debugging access rules.
      tags: [Channel Rules]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groups:
                  type: array
                  items:
                    type: string
                  description: JWT groups to test access for
              required:
                - groups
      responses:
        "200":
          description: Allowed patterns
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed_patterns:
                    type: array
                    items:
                      type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin JWT (Ed25519 keypair, iss:"sukko-admin") or tenant JWT (per-tenant asymmetric keys).
        Admin JWT is checked first; if the issuer is not "sukko-admin", tenant JWT validation is used.
  parameters:
    tenantSlug:
      name: tenantSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]{2,62}$"
      description: Tenant slug — the human-readable Kafka namespace and URL identifier (3–63 chars, lowercase alphanumeric + hyphens, must start with a letter)
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code (UPPER_SNAKE_CASE)
          examples: ["TENANT_NOT_FOUND"]
        message:
          type: string
          description: Human-readable error message
          examples: ["tenant not found"]
...
    TenantChannelRules:
      type: object
      properties:
        tenant_id:
          type: string
        rules:
          $ref: "#/components/schemas/ChannelRules"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChannelRules:
      type: object
      properties:
        public:
          type: array
          items:
            type: string
          description: "Channel patterns accessible to all authenticated users (e.g., '*.messages')"
        group_mappings:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: "Group name → channel patterns mapping"
        default:
          type: array
          items:
            type: string
          description: Fallback channel patterns when no group matches
        publish_public:
          type: array
          items:
            type: string
          description: Channel patterns anyone can publish to
        publish_group_mappings:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: "Group name → publish channel patterns"
        publish_default:
          type: array
          items:
            type: string
          description: Fallback publish channel patterns
    SetChannelRulesRequest:
      type: object
      properties:
        public:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 256
        group_mappings:
          type: object
          additionalProperties:
            type: array
            maxItems: 20
            items:
              type: string
              maxLength: 256
        default:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 256
        publish_public:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 256
        publish_group_mappings:
          type: object
          additionalProperties:
            type: array
            maxItems: 20
            items:
              type: string
              maxLength: 256
        publish_default:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 256
...